[Refactor] encapsulate into `BlockInsertPalette` for quick document block insertion
